fix23

+++ src/main/java/com/fuib/lotus/agents/params/values/WorkTimeDiffDaysValue.java
@@ -14,8 +14,8 @@ public class WorkTimeDiffDaysValue extends WorkTimeDiffMinuteValue {
     @Override
     public Vector getColumnValue(Document doc) throws NotesException {
         Vector v = super.getColumnValue(doc);
-        double workDays = ((Integer) v.get(0)) / 480;    // 60 min per 1 hour and 8 working hours a day
-        Vector<Double> vr = new Vector<Double>(1);
+        Long workDays = ((Long) v.get(0)) / 480;    // 60 min per 1 hour and 8 working hours a day
+        Vector<Long> vr = new Vector<Long>(1);
         vr.add(workDays);
         return vr;
     }
